refactor validateShift and calculateCoverage from Scheduler to utils
